+++ CHANGELOG.md
@@ -15,6 +15,8 @@ FlatLaf Change Log
 - ToolTip: Do not show empty tooltip component if tooltip text is an empty
   string. (issue #134)
 - Button: Support specifying button border width.
+- ComboBox: Fixed width of popup, which was too small if popup is wider than
+  combo box and vertical scroll bar is visible. (issue #137)
 - ComboBox: Changed maximum row count of popup list to 15 (was 20). Set UI value
   `ComboBox.maximumRowCount` to any integer to use a different value.